[Decomp][Spell] Cohesion split: extract area/raid targeting to SpellAreaTargets.cpp (#207)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  helper:PrioritizeManaUnitWraper,helper:PrioritizeMana,helper:PrioritizeHealthUnitWraper,helper:PrioritizeHealth,Spell::FillRaidOrPartyManaPriorityTargets,Spell::FillRaidOrPartyHealthPriorityTargets
helper:PrioritizeManaUnitWraper,helper:PrioritizeMana,helper:PrioritizeHealthUnitWraper,helper:PrioritizeHealth,Spell::FillRaidOrPartyManaPriorityTargets,Spell::FillRaidOrPartyHealthPriorityTargets
Ported-from: mangosthree/server@b97ac06eab53
