+++ src/game/WorldHandlers/Spell.cpp
@@ -1057,90 +1057,6 @@ SpellCastResult Spell::CanOpenLock(SpellEffectIndex effIndex, uint32 lockId, Ski
 }
 
 /**
- * Fill target list by units around (x,y) points at radius distance
-
- * @param targetUnitMap        Reference to target list that filled by function
- * @param radius               Radius around (x,y) for target search
- * @param pushType             Additional rules for target area selection (in front, angle, etc)
- * @param spellTargets         Additional rules for target selection base at hostile/friendly state to original spell caster
- */
-void Spell::FillAreaTargets(UnitList& targetUnitMap, float radius, SpellNotifyPushType pushType, SpellTargets spellTargets, WorldObject* originalCaster /*=NULL*/)
-{
-    MaNGOS::SpellNotifierCreatureAndPlayer notifier(*this, targetUnitMap, radius, pushType, spellTargets, originalCaster);
-    Cell::VisitAllObjects(notifier.GetCenterX(), notifier.GetCenterY(), m_caster->GetMap(), notifier, radius);
-}
-
-/**
- * @brief Fills a target list with party or raid members around a reference unit.
- *
- * @param targetUnitMap The target list being populated.
- * @param member The reference member.
- * @param radius The search radius.
- * @param raid True to include the whole raid; false to limit to the subgroup.
- * @param withPets True to include pets.
- * @param withcaster True to include the caster when applicable.
- */
-void Spell::FillRaidOrPartyTargets(UnitList& targetUnitMap, Unit* member, float radius, bool raid, bool withPets, bool withcaster)
-{
-    Player* pMember = member->GetCharmerOrOwnerPlayerOrPlayerItself();
-    Group* pGroup = pMember ? pMember->GetGroup() : NULL;
-
-    if (pGroup)
-    {
-        uint8 subgroup = pMember->GetSubGroup();
-
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* Target = itr->getSource();
-
-            // IsHostileTo check duel and controlled by enemy
-            if (Target && (raid || subgroup == Target->GetSubGroup())
-                && !m_caster->IsHostileTo(Target))
-            {
-                if ((Target == m_caster && withcaster) ||
-                    (Target != m_caster && m_caster->IsWithinDistInMap(Target, radius)))
-                {
-                    targetUnitMap.push_back(Target);
-                }
-
-                if (withPets)
-                {
-                    if (Pet* pet = Target->GetPet())
-                    {
-                        if ((pet == m_caster && withcaster) ||
-                            (pet != m_caster && m_caster->IsWithinDistInMap(pet, radius)))
-                        {
-                            targetUnitMap.push_back(pet);
-                        }
-                    }
-                }
-            }
-        }
-    }
-    else
-    {
-        Unit* ownerOrSelf = pMember ? pMember : member->GetCharmerOrOwnerOrSelf();
-        if ((ownerOrSelf == m_caster && withcaster) ||
-            (ownerOrSelf != m_caster && m_caster->IsWithinDistInMap(ownerOrSelf, radius)))
-        {
-            targetUnitMap.push_back(ownerOrSelf);
-        }
-
-        if (withPets)
-        {
-            if (Pet* pet = ownerOrSelf->GetPet())
-            {
-                if ((pet == m_caster && withcaster) ||
-                    (pet != m_caster && m_caster->IsWithinDistInMap(pet, radius)))
-                {
-                    targetUnitMap.push_back(pet);
-                }
-            }
-        }
-    }
-}
-
-/**
  * @brief Gets the world object that should be used as the effective spell origin.
  *
  * @return The effective caster world object.
