+++ src/game/WorldHandlers/SpellAreaTargets.cpp
@@ -1,0 +1,148 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file SpellAreaTargets.cpp
+ * @brief Cohesion split of Spell.cpp.
+ *        Re-applied onto MangosOne TBC 2.4.3; same class, pure code move,
+ *        no behaviour change. CMake file(GLOB) picks this TU up automatically.
+ */
+
+#include "Spell.h"
+#include "Database/DatabaseEnv.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "SpellMgr.h"
+#include "Player.h"
+#include "Pet.h"
+#include "Unit.h"
+#include "DynamicObject.h"
+#include "Group.h"
+#include "UpdateData.h"
+#include "ObjectAccessor.h"
+#include "CellImpl.h"
+#include "Policies/Singleton.h"
+#include "SharedDefines.h"
+#include "LootMgr.h"
+#include "VMapFactory.h"
+#include "BattleGround/BattleGround.h"
+#include "Util.h"
+#include "Chat.h"
+#include "SQLStorages.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+extern pEffect SpellEffects[TOTAL_SPELL_EFFECTS];
+
+/**
+ * Fill target list by units around (x,y) points at radius distance
+
+ * @param targetUnitMap        Reference to target list that filled by function
+ * @param radius               Radius around (x,y) for target search
+ * @param pushType             Additional rules for target area selection (in front, angle, etc)
+ * @param spellTargets         Additional rules for target selection base at hostile/friendly state to original spell caster
+ */
+void Spell::FillAreaTargets(UnitList& targetUnitMap, float radius, SpellNotifyPushType pushType, SpellTargets spellTargets, WorldObject* originalCaster /*=NULL*/)
+{
+    MaNGOS::SpellNotifierCreatureAndPlayer notifier(*this, targetUnitMap, radius, pushType, spellTargets, originalCaster);
+    Cell::VisitAllObjects(notifier.GetCenterX(), notifier.GetCenterY(), m_caster->GetMap(), notifier, radius);
+}
+
+/**
+ * @brief Fills a target list with party or raid members around a reference unit.
+ *
+ * @param targetUnitMap The target list being populated.
+ * @param member The reference member.
+ * @param radius The search radius.
+ * @param raid True to include the whole raid; false to limit to the subgroup.
+ * @param withPets True to include pets.
+ * @param withcaster True to include the caster when applicable.
+ */
+void Spell::FillRaidOrPartyTargets(UnitList& targetUnitMap, Unit* member, float radius, bool raid, bool withPets, bool withcaster)
+{
+    Player* pMember = member->GetCharmerOrOwnerPlayerOrPlayerItself();
+    Group* pGroup = pMember ? pMember->GetGroup() : NULL;
+
+    if (pGroup)
+    {
+        uint8 subgroup = pMember->GetSubGroup();
+
+        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
+        {
+            Player* Target = itr->getSource();
+
+            // IsHostileTo check duel and controlled by enemy
+            if (Target && (raid || subgroup == Target->GetSubGroup())
+                && !m_caster->IsHostileTo(Target))
+            {
+                if ((Target == m_caster && withcaster) ||
+                    (Target != m_caster && m_caster->IsWithinDistInMap(Target, radius)))
+                {
+                    targetUnitMap.push_back(Target);
+                }
+
+                if (withPets)
+                {
+                    if (Pet* pet = Target->GetPet())
+                    {
+                        if ((pet == m_caster && withcaster) ||
+                            (pet != m_caster && m_caster->IsWithinDistInMap(pet, radius)))
+                        {
+                            targetUnitMap.push_back(pet);
+                        }
+                    }
+                }
+            }
+        }
+    }
+    else
+    {
+        Unit* ownerOrSelf = pMember ? pMember : member->GetCharmerOrOwnerOrSelf();
+        if ((ownerOrSelf == m_caster && withcaster) ||
+            (ownerOrSelf != m_caster && m_caster->IsWithinDistInMap(ownerOrSelf, radius)))
+        {
+            targetUnitMap.push_back(ownerOrSelf);
+        }
+
+        if (withPets)
+        {
+            if (Pet* pet = ownerOrSelf->GetPet())
+            {
+                if ((pet == m_caster && withcaster) ||
+                    (pet != m_caster && m_caster->IsWithinDistInMap(pet, radius)))
+                {
+                    targetUnitMap.push_back(pet);
+                }
+            }
+        }
+    }
+}
